feat: 이미지 업로드 기능을 적용합니다.
