--- src/components/mypage/edit.tsx
@@ -1,149 +1,0 @@
-import React, { useState, useEffect, useRef } from "react";
-import { Box, Image, Input, Button, Text } from "@chakra-ui/react";
-import useUpload from "@hooks/useUpload";
-
-type Props = {};
-
-const Edit: React.FC<Props> = ({}) => {
-	const [upload, file] = useUpload();
-	const [imageSrc, setImageSrc] = useState("");
-
-	const inputRef = useRef(null);
-	const getBase64 = (file: File) => new Promise((resolve, reject) => {
-		if (file === null) {
-			reject()
-		}
-
-		var reader = new FileReader();
-		reader.readAsDataURL(file);
-		reader.onload = function () {
-			resolve(reader.result);
-		};
-		reader.onerror = function (error) {
-			reject(error)
-		};
-	})
-
-	useEffect(() => {
-		if (file) {
-			getBase64(file).then((src: any) => {
-				setImageSrc(src)
-			}).catch(error => {
-				console.error(error)
-			})
-		} else {
-			setImageSrc('https://bit.ly/dan-abramov')
-		}
-	}, [file]);
-
-	return (
-		<Box>
-			<Box position='absolute' w='1920px' h='56px'>
-				<Text fontSize='20px'>COQUALITY (상단바)</Text>
-			</Box>
-			<Box></Box>
-			<Text position='absolute' mt='104px' h='48px' fontSize='32px' fontWeight='600'>
-				프로필 수정
-			</Text>
-			<Box position='absolute' display='flex' mt='200px' w='1080px' h='120px'>
-				<Image
-					w='120px'
-					h='120px'
-					borderRadius='100%'
-					ref={inputRef}
-					src={imageSrc}
-					alt='Profile Image'
-				/>
-				<Box display='flex' flexDirection='column' ml='20px'>
-					<Button
-						mb='12px'
-						h='44px'
-						backgroundColor='black'
-						color='white'
-						borderRadius='50px'
-						p='10px 30px'
-						fontSize='16px'
-						fontWeight='600'
-						onClick={upload}
-					>
-						이미지 업로드
-					</Button>
-					<Button
-						colorScheme='black'
-						variant='outline'
-						h='44px'
-						fontSize='16px'
-						fontWeight='600'
-						borderRadius='50px'
-						p='12px 30px'
-						onClick={() => setImageSrc('')}
-					>
-						이미지 제거
-					</Button>
-				</Box>
-			</Box>
-			<Box position='absolute' w='1080px' h='73px' gap='16px' mt='400px'>
-				<Text fontSize='14px' fontWeight='600' color='black' letterSpacing='-0.01em'>
-					이메일
-				</Text>
-				<Input
-					mt='16px'
-					p='0px 0px 12px'
-					gap='10px'
-					w='1080px'
-					h='36px'
-					border='0px'
-					borderBottom='1px solid #DDDDDD'
-					placeholder='abcdefgh@gmail.com'
-					color='#999999'
-					fontSize='16px'
-					fontWeight='400'
-				></Input>
-				<Text mt='32px' fontSize='14px' fontWeight='600' color='black' letterSpacing='-0.01em'>
-					닉네임
-				</Text>
-				<Input
-					mt='16px'
-					p='0px 0px 12px'
-					gap='10px'
-					w='1080px'
-					h='36px'
-					border='0px'
-					borderBottom='1px solid #DDDDDD'
-				></Input>
-				<Text fontSize='14px' mt='32px' fontWeight='600' color='black' letterSpacing='-0.01em'>
-					소개
-				</Text>
-				<Input
-					mt='16px'
-					p='0px 0px 12px'
-					gap='10px'
-					w='1080px'
-					h='36px'
-					border='0px'
-					borderBottom='1px solid #DDDDDD'
-					placeholder='나를 설명하는 한 줄 소개를 작성해 보세요'
-					color='#999999'
-					fontSize='16px'
-					fontWeight='400'
-				></Input>
-			</Box>
-			<Button
-				position='absolute'
-				mt='793px'
-				backgroundColor='#999999'
-				color='white'
-				borderRadius='50px'
-				p='10px 30px'
-				h='44px'
-				w='255px'
-				fontSize='16px'
-				fontWeight='600'
-			>
-				수정완료
-			</Button>
-		</Box>
-	);
-};
-
-export default Edit;
