+++ src/containers/ProfileEdit/index.tsx
@@ -1,24 +1,58 @@
-import React from "react"
+import React, { useState, useRef, useEffect } from "react"
 import useUpload from "@hooks/useUpload"
 import styled from "@emotion/styled"
 import { colors } from "@constants/colors"
 import Button from "@components/inputs/Button"
 import Textfield from "@components/inputs/Textfield"
 import Link from "next/link"
+import Image from "next/image"
 
 type Props = {}
 
 const ProfileEdit: React.FC<Props> = ({}) => {
+  const [upload, file] = useUpload();
+	const [imageSrc, setImageSrc] = useState("");
+
+	const inputRef = useRef(null);
+	const getBase64 = (file: File) => new Promise((resolve, reject) => {
+		if (file === null) {
+			reject()
+		}
+
+		var reader = new FileReader();
+		reader.readAsDataURL(file);
+		reader.onload = function () {
+			resolve(reader.result);
+		};
+		reader.onerror = function (error) {
+			reject(error)
+		};
+	})
+
+	useEffect(() => {
+		if (file) {
+			getBase64(file).then((src: any) => {
+				setImageSrc(src)
+			}).catch(error => {
+				console.error(error)
+			})
+		}  
+	}, [file]);
+
   return (
     <StyledWrapper className="common-container">
       <div className="header common-h1-sb">프로필 수정</div>
       <div className="profile-info">
-        <div className="lt"></div>
+        <img 
+					className="lt"
+					src={imageSrc}
+          ref={inputRef}
+          alt='' />
         <div className="rt">
-          <Button className="img-btn upload-btn common-h5-sb">
+          <Button className="img-btn upload-btn common-h5-sb" onClick={upload}>
             이미지 업로드
           </Button>
-          <Button className="img-btn common-h5-sb" variant="outline">
+          <Button className="img-btn common-h5-sb" variant="outline" onClick={() => setImageSrc('')}>
             이미지 삭제
           </Button>
         </div>
